pyproject.toml
[project]
name = "giantsmind"
version = "0.1.0"
description = "A Python package that provides tools for interacting with scientific articles PDFs."
authors = [{name = "Pierre Enel", email = "pierre.enel@gmail.com"}]

dependencies = [
    "langchain",
    "langchain_community",
    "llama_parse",
    "chromadb",
    "sqlalchemy",
    "pymupdf",
    "PyPDF2",
    "requests",
    "python-dotenv>=0.19.0",
    "platformdirs",
]

classifiers = [
   "Development Status :: 2 - Pre-Alpha",
   "Intended Audience :: Science/Research",
   "License :: OSI Approved :: BSD License",
   'Operating System :: Unix',
   "Programming Language :: Python :: 3.12",
   "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[tool.black]
line-length = 110

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = "-vv"
testpaths = ["tests"]

[project.scripts]
giantsmind = "giantsmind.cli_entry:main"

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "isort",
    "pytest",
    "pytest-mock",
]